Fixed export progress bar
