+++ Core/Application/Application.csproj
@@ -25,6 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\ApsimX\APSIM.Shared\APSIM.Shared.csproj" />
     <ProjectReference Include="..\..\..\ApsimX\Models\Models.csproj" />
     <ProjectReference Include="..\Domain\Domain.csproj" />
   </ItemGroup>
