+++ Core/Application/CQRS/PlotData/InsertPlotDataTableCommand.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using MediatR;
+using Rems.Application.Common;
 using Rems.Application.Common.Extensions;
 using Rems.Application.Common.Interfaces;
 using Rems.Domain.Entities;
@@ -13,7 +14,7 @@ using Unit = MediatR.Unit;
 
 namespace Rems.Application.CQRS
 {
-    public class InsertPlotDataTableCommand : IRequest<Unit>
+    public class InsertPlotDataTableCommand : IRequest
     {
         public DataTable Table { get; set; }
 
@@ -22,7 +23,7 @@ namespace Rems.Application.CQRS
         public string Type { get; set; }
     }
 
-    public class InsertPlotDataTableCommandHandler : IRequestHandler<InsertPlotDataTableCommand, Unit>
+    public class InsertPlotDataTableCommandHandler : IRequestHandler<InsertPlotDataTableCommand>
     {
         private readonly IRemsDbContext _context;
 
@@ -70,7 +71,7 @@ namespace Rems.Application.CQRS
                     _context.Add(data);
                 }
 
-                EventManager.InvokeProgressIncremented(null, EventArgs.Empty);
+                EventManager.InvokeProgressIncremented();
             }
             _context.SaveChanges();
