+++ Core/Application/CQRS/SoilLayerData/InsertSoilLayerTableCommand.cs
@@ -5,6 +5,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using MediatR;
+using Rems.Application.Common;
 using Rems.Application.Common.Extensions;
 using Rems.Application.Common.Interfaces;
 using Rems.Domain.Entities;
@@ -13,7 +14,7 @@ using Unit = MediatR.Unit;
 
 namespace Rems.Application.CQRS
 {
-    public class InsertSoilLayerTableCommand : IRequest<Unit>
+    public class InsertSoilLayerTableCommand : IRequest
     {
         public DataTable Table { get; set; }
 
@@ -22,7 +23,7 @@ namespace Rems.Application.CQRS
         public string Type { get; set; }
     }
 
-    public class InsertSoilLayerTableCommandHandler : IRequestHandler<InsertSoilLayerTableCommand, Unit>
+    public class InsertSoilLayerTableCommandHandler : IRequestHandler<InsertSoilLayerTableCommand>
     {
         private readonly IRemsDbContext _context;
 
@@ -63,7 +64,7 @@ namespace Rems.Application.CQRS
                     };
                     _context.Attach(data);
                 }
-                EventManager.InvokeProgressIncremented(null, EventArgs.Empty);
+                EventManager.InvokeProgressIncremented();
             }
             _context.SaveChanges();
