+++ Core/Application/CQRS/Tables/InsertOperationsTableCommand.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using Rems.Application.Common;
 using Rems.Application.Common.Extensions;
 using Rems.Application.Common.Interfaces;
 using Rems.Domain.Entities;
@@ -15,14 +16,14 @@ using Unit = MediatR.Unit;
 
 namespace Rems.Application.CQRS
 {
-    public class InsertOperationsTableCommand : IRequest<Unit>
+    public class InsertOperationsTableCommand : IRequest
     {
         public DataTable Table { get; set; }
 
         public Type Type { get; set; }
     }
 
-    public class InsertOperationsTableCommandHandler : IRequestHandler<InsertOperationsTableCommand, Unit>
+    public class InsertOperationsTableCommandHandler : IRequestHandler<InsertOperationsTableCommand>
     {
         private readonly IRemsDbContext _context;
 
@@ -77,7 +78,7 @@ namespace Rems.Application.CQRS
                     _context.Attach(result);
                 }
 
-                EventManager.InvokeProgressIncremented(null, EventArgs.Empty);
+                EventManager.InvokeProgressIncremented();
             }
 
             _context.SaveChanges();
