+++ Core/Application/CQRS/Tables/InsertTraitTableCommand.cs
@@ -7,6 +7,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using MediatR;
+using Rems.Application.Common;
 using Rems.Application.Common.Extensions;
 using Rems.Application.Common.Interfaces;
 using Rems.Domain.Entities;
@@ -15,7 +16,7 @@ using Unit = MediatR.Unit;
 
 namespace Rems.Application.CQRS
 {
-    public class InsertTraitTableCommand : IRequest<Unit>
+    public class InsertTraitTableCommand : IRequest
     {
         public DataTable Table { get; set; }
 
@@ -24,7 +25,7 @@ namespace Rems.Application.CQRS
         public Type Dependency { get; set; }
     }
 
-    public class InsertTraitTableCommandHandler : IRequestHandler<InsertTraitTableCommand, Unit>
+    public class InsertTraitTableCommandHandler : IRequestHandler<InsertTraitTableCommand>
     {
         private readonly IRemsDbContext _context;
 
@@ -79,7 +80,7 @@ namespace Rems.Application.CQRS
 
                 _context.Attach(entity);
 
-                EventManager.InvokeProgressIncremented(null, EventArgs.Empty);
+                EventManager.InvokeProgressIncremented();
             }
             _context.SaveChanges();
