+++ Infrastructure/Infrastructure/Infrastructure.csproj
@@ -51,6 +51,15 @@
     <None Update="DataFiles\apsimx\Sorghum.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="DataFiles\xlsx\Demo_1_Information.xlsx">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="DataFiles\xlsx\Demo_2_Experiments.xlsx">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="DataFiles\xlsx\Demo_3_Data.xlsx">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
